+++ evals/__init__.py
@@ -1,11 +1,8 @@
-from .api import (
-    CompletionFn,
-    CompletionResult,
+from .api import CompletionFn, CompletionResult, DummyCompletionFn, record_and_check_match
+from .completion_fns.openai import (
     OpenAIChatCompletionFn,
     OpenAICompletionFn,
     OpenAICompletionResult,
-    record_and_check_match,
-    sample_freeform,
 )
 from .data import get_csv, get_json, get_jsonl, get_jsonls, get_lines, iter_jsonls
 from .eval import Eval
