+++ evals/api.py
@@ -7,18 +7,8 @@ import logging
 from abc import ABC, abstractmethod
 from typing import Any, Callable, Optional, Protocol, Union, runtime_checkable
 
-from evals.prompt.base import (
-    ChatCompletionPrompt,
-    CompletionPrompt,
-    OpenAICreateChatPrompt,
-    OpenAICreatePrompt,
-    Prompt,
-)
-from evals.record import record_match, record_sampling
-from evals.utils.api_utils import (
-    openai_chat_completion_create_retrying,
-    openai_completion_create_retrying,
-)
+from evals.prompt.base import OpenAICreateChatPrompt, OpenAICreatePrompt, Prompt
+from evals.record import record_match
 
 logger = logging.getLogger(__name__)
 
@@ -33,7 +23,7 @@ class CompletionResult(ABC):
 class CompletionFn(Protocol):
     def __call__(
         self,
-        prompt: Union[OpenAICreatePrompt, OpenAICreateChatPrompt, Prompt],
+        prompt: Union[str, OpenAICreateChatPrompt],
         **kwargs,
     ) -> CompletionResult:
         """
@@ -48,127 +38,6 @@ class CompletionFn(Protocol):
         The result of the API call.
         The prompt that was fed into the API call as a str.
         """
-
-
-class OpenAICompletionResult(CompletionResult):
-    def __init__(self, raw_data: Any, prompt: Any):
-        self.raw_data = raw_data
-        self.prompt = prompt
-
-    def get_completions(self) -> list[str]:
-        raise NotImplementedError
-
-
-class OpenAIChatCompletionResult(OpenAICompletionResult):
-    def get_completions(self) -> list[str]:
-        completions = []
-        if self.raw_data and "choices" in self.raw_data:
-            for choice in self.raw_data["choices"]:
-                if "message" in choice:
-                    completions.append(choice["message"]["content"])
-        return completions
-
-
-class OpenAICompletionResult(OpenAICompletionResult):
-    def get_completions(self) -> list[str]:
-        completions = []
-        if self.raw_data and "choices" in self.raw_data:
-            for choice in self.raw_data["choices"]:
-                if "text" in choice:
-                    completions.append(choice["text"])
-        return completions
-
-
-class OpenAICompletionFn(CompletionFn):
-    def __init__(
-        self,
-        model: Optional[str] = None,
-        api_base: Optional[str] = None,
-        api_key: Optional[str] = None,
-        n_ctx: Optional[int] = None,
-        extra_options: Optional[dict] = {},
-    ):
-        self.model = model
-        self.api_base = api_base
-        self.api_key = api_key
-        self.n_ctx = n_ctx
-        self.extra_options = extra_options
-
-    def __call__(
-        self,
-        prompt: Union[OpenAICreatePrompt, Prompt],
-        **kwargs,
-    ) -> OpenAICompletionResult:
-        if not isinstance(prompt, Prompt):
-            assert (
-                isinstance(prompt, str)
-                or (isinstance(prompt, list) and all(isinstance(token, int) for token in prompt))
-                or (isinstance(prompt, list) and all(isinstance(token, str) for token in prompt))
-                or (isinstance(prompt, list) and all(isinstance(msg, dict) for msg in prompt))
-            ), f"Got type {type(prompt)}, with val {type(prompt[0])} for prompt, expected str or list[int] or list[str] or list[dict[str, str]]"
-
-            prompt = CompletionPrompt(
-                raw_prompt=prompt,
-            )
-
-        openai_create_prompt: OpenAICreatePrompt = prompt.to_formatted_prompt()
-
-        result = openai_completion_create_retrying(
-            model=self.model,
-            api_base=self.api_base,
-            api_key=self.api_key,
-            prompt=openai_create_prompt,
-            **{**kwargs, **self.extra_options},
-        )
-        result = OpenAICompletionResult(raw_data=result, prompt=openai_create_prompt)
-        record_sampling(prompt=result.prompt, sampled=result.get_completions())
-        return result
-
-
-class OpenAIChatCompletionFn(CompletionFn):
-    def __init__(
-        self,
-        model: Optional[str] = None,
-        api_base: Optional[str] = None,
-        api_key: Optional[str] = None,
-        n_ctx: Optional[int] = None,
-        extra_options: Optional[dict] = {},
-    ):
-        self.model = model
-        self.api_base = api_base
-        self.api_key = api_key
-        self.n_ctx = n_ctx
-        self.extra_options = extra_options
-
-    def __call__(
-        self,
-        prompt: Union[OpenAICreateChatPrompt, Prompt],
-        **kwargs,
-    ) -> OpenAIChatCompletionResult:
-        if not isinstance(prompt, Prompt):
-            assert (
-                isinstance(prompt, str)
-                or (isinstance(prompt, list) and all(isinstance(token, int) for token in prompt))
-                or (isinstance(prompt, list) and all(isinstance(token, str) for token in prompt))
-                or (isinstance(prompt, list) and all(isinstance(msg, dict) for msg in prompt))
-            ), f"Got type {type(prompt)}, with val {type(prompt[0])} for prompt, expected str or list[int] or list[str] or list[dict[str, str]]"
-
-            prompt = ChatCompletionPrompt(
-                raw_prompt=prompt,
-            )
-
-        openai_create_prompt: OpenAICreateChatPrompt = prompt.to_formatted_prompt()
-
-        result = openai_chat_completion_create_retrying(
-            model=self.model,
-            api_base=self.api_base,
-            api_key=self.api_key,
-            messages=openai_create_prompt,
-            **{**kwargs, **self.extra_options},
-        )
-        result = OpenAIChatCompletionResult(raw_data=result, prompt=openai_create_prompt)
-        record_sampling(prompt=result.prompt, sampled=result.get_completions())
-        return result
 
 
 class DummyCompletionResult(CompletionResult):
@@ -190,6 +59,19 @@ def record_and_check_match(
     separator: Callable[[str], bool] = None,
     options: Optional[list[str]] = None,
 ):
+    """
+    Records and checks if a sampled response from a CompletionFn matches the expected result.
+
+    Args:
+        prompt: The input prompt.
+        sampled: The sampled response from the model.
+        expected: The expected response or list of responses.
+        separator: Optional function to check if a character is a separator.
+        options: Optional list of options to match against the sampled response.
+
+    Returns:
+        The matched option or None if no match found.
+    """
     if isinstance(expected, tuple):
         expected = list(expected)
     elif not isinstance(expected, list):
@@ -221,48 +103,3 @@ def record_and_check_match(
     result["match"] = match
     record_match(match, expected=expected, picked=picked, sampled=sampled, options=options)
     return picked
-
-
-def sample_freeform(
-    prompt: Union[OpenAICreatePrompt, OpenAICreateChatPrompt, Prompt],
-    *,
-    completion_fn: CompletionFn = OpenAIChatCompletionFn(),
-    temperature: float = 1.0,
-    top_p: float = 0.9,
-    max_tokens: int = 512,
-    stop: Optional[str] = None,
-    n_samples: Optional[int] = None,
-    **kwargs,
-) -> Union[str, list[str], dict]:
-    """
-    Samples a freeform response from the specified model, records the sampling,
-        and returns the sampled text.
-
-    ARGS
-    ====
-    `model_spec`: See `completion_query`.
-    `prompt`: See `completion_query`.
-    `temperature`: Passed to `openai.Completion.create`.
-    `top_p`: Passed to `openai.Completion.create`.
-    `max_tokens`: Passed to `openai.Completion.create`.
-    `stop`: Passed to `openai.Completion.create`.
-    `n_samples`: The number of samples to generate (1 if None).
-    `kwargs`: See `completion_query`.
-
-    RETURNS
-    =======
-    Returns the sampled text, or a list of sampled texts if
-        `n_samples` is not None.
-    """
-    result = completion_fn(
-        prompt=prompt,
-        temperature=temperature,
-        top_p=top_p,
-        max_tokens=max_tokens,
-        stop=stop,
-        n=(1 if n_samples is None else n_samples),
-        headers={},
-        **kwargs,
-    )
-    sampled = result.get_completions()[0]
-    return sampled
